Add CatImage
a CatImage fetches an XML response from the thecatapi.com and exposes a
source_url (thecatapi.com, for retribution to the copyright holder) and
a url (we're gonna need that to fetch the image for the save to disk
part).
a parser can be specified on a CatImage instance as long as it returns
something exposing a url and source_url.