Add the CatFetcher
took a break for diner and came back to it.

add a CatFetcher which will fetch a given URL and write the response to
a file at the given path.
updates the Cat#save method to returns a friendly message if everything
went fine, or not. this will be written to the console.